Provide no-op macros in Solana targets
