+++ codama-macros/src/lib.rs
@@ -1,70 +1,63 @@
-use codama_attributes::{AttributeContext, CodamaAttribute};
-use codama_errors::{CodamaError, CodamaResult};
-use codama_koroks::CrateKorok;
-use codama_stores::CrateStore;
 use proc_macro::TokenStream;
-use proc_macro2::TokenStream as TokenStream2;
-use quote::quote;
+#[cfg(not(target_os = "solana"))]
+mod attributes;
+#[cfg(not(target_os = "solana"))]
+mod derives;
+
+fn codama_derive(input: TokenStream) -> TokenStream {
+    #[cfg(not(target_os = "solana"))]
+    {
+        derives::codama_derive(input)
+    }
+    #[cfg(target_os = "solana")]
+    {
+        input
+    }
+}
+
+fn codama_attribute(attr: TokenStream, input: TokenStream) -> TokenStream {
+    #[cfg(not(target_os = "solana"))]
+    {
+        attributes::codama_attribute(attr, input)
+    }
+    #[cfg(target_os = "solana")]
+    {
+        let _ = attr;
+        input
+    }
+}
 
 #[proc_macro_derive(CodamaAccount, attributes(codama))]
 pub fn codama_account_derive(input: TokenStream) -> TokenStream {
-    codama_derive(input.into())
-        .unwrap_or_else(CodamaError::into_compile_error)
-        .into()
+    codama_derive(input)
 }
 
 #[proc_macro_derive(CodamaAccounts, attributes(codama))]
 pub fn codama_accounts_derive(input: TokenStream) -> TokenStream {
-    codama_derive(input.into())
-        .unwrap_or_else(CodamaError::into_compile_error)
-        .into()
+    codama_derive(input)
 }
 
 #[proc_macro_derive(CodamaErrors, attributes(codama))]
 pub fn codama_errors_derive(input: TokenStream) -> TokenStream {
-    codama_derive(input.into())
-        .unwrap_or_else(CodamaError::into_compile_error)
-        .into()
+    codama_derive(input)
 }
 
 #[proc_macro_derive(CodamaInstruction, attributes(codama))]
 pub fn codama_instruction_derive(input: TokenStream) -> TokenStream {
-    codama_derive(input.into())
-        .unwrap_or_else(CodamaError::into_compile_error)
-        .into()
+    codama_derive(input)
 }
 
 #[proc_macro_derive(CodamaInstructions, attributes(codama))]
 pub fn codama_instructions_derive(input: TokenStream) -> TokenStream {
-    codama_derive(input.into())
-        .unwrap_or_else(CodamaError::into_compile_error)
-        .into()
+    codama_derive(input)
 }
 
 #[proc_macro_derive(CodamaType, attributes(codama))]
 pub fn codama_type_derive(input: TokenStream) -> TokenStream {
-    codama_derive(input.into())
-        .unwrap_or_else(CodamaError::into_compile_error)
-        .into()
-}
-
-fn codama_derive(input: TokenStream2) -> CodamaResult<TokenStream2> {
-    let store = CrateStore::hydrate(input)?;
-    CrateKorok::parse(&store)?;
-    Ok(quote! {})
+    codama_derive(input)
 }
 
 #[proc_macro_attribute]
 pub fn codama(attr: TokenStream, input: TokenStream) -> TokenStream {
-    codama_attribute(attr.into(), input.into())
-        .unwrap_or_else(CodamaError::into_compile_error)
-        .into()
-}
-
-fn codama_attribute(attr: TokenStream2, input: TokenStream2) -> CodamaResult<TokenStream2> {
-    let attr: syn::Attribute = syn::parse_quote! { #[codama(#attr)] };
-    let item: syn::Item = syn::parse2(input.clone())?;
-    let ctx: AttributeContext = (&item).into();
-    CodamaAttribute::parse(&attr, &ctx)?;
-    Ok(input)
+    codama_attribute(attr, input)
 }
